Gradle 사용하도록 변경
intellij 설정에서 maven 사용해 lib 추가 -> gradle dependencies 추가로 변경
