+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Mon Mar 04 15:12:52 KST 2024
+gradle.version=7.5.1
